Made allocator spiteful against jerlim
Allocates 2x the number of total cores on the machine if jerlim is
dominating the machine and nobody else is on it. He is spamming the
cluster with processes, one process for every single core on every
non-reserved machine, depriving everyone else from running processes at
reasonable speeds. This commit aims to increase the cost for him to run
his processes in this manner.